refactor(card-browser): move Preview to Fragment
